Make validation tests work
