+++ test.sh
@@ -4,6 +4,6 @@ set -eux
 
 cd $(dirname $0)
 
-time cargo test
+time cargo test --all
 
 cd -
